+++ config/config-passport.js
@@ -1,6 +1,6 @@
 const passport = require("passport");
 const passportJWT = require("passport-jwt");
-const User = require("../service/schemas/users");
+const {User} = require("../models/users");
 require("dotenv").config();
 const secret = process.env.SECRET;
