+++ controllers/users/index.js
@@ -1,0 +1,12 @@
+const signupCtrl = require("./singup");
+const loginCtrl = require("./login");
+const logoutCtrl = require("./logout");
+const currentCtrl = require("./getCurrent");
+
+module.exports = {
+    signupCtrl,
+    loginCtrl,
+    logoutCtrl,
+    currentCtrl
+}
+
