+++ controllers/users/logout.js
@@ -1,0 +1,11 @@
+const {User} = require("../../models/index");
+
+const logoutCtrl = async(req, res)=> {
+    const {_id} = req.user;
+    await User.findByIdAndUpdate(_id, {token: null});
+    res.status(204).json();
+}
+
+module.exports = logoutCtrl;
+
+
