--- middlewares/auth.js
@@ -1,19 +1,0 @@
-const passport = require("passport");
-
-const auth = (req, res, next) => {
-    passport.authenticate("jwt", {session: false}, (err, user) => {
-        if (!user || err ) {
-            return res.status(401).json({
-                status: "error",
-                code: 401,
-                message: "Not authorized",
-                data: "Unauthorized",
-            });
-        }
-        req.user = user;
-        next();
-    })(req, res, next);
-};
-
-module.exports = auth;
-
