+++ middlewares/index.js
@@ -1,0 +1,9 @@
+const authToken = require("./authToken");
+const ctrlWrapper = require("./ctrlWrapper");
+const validateBody = require("./validateBody");
+
+module.exports = {
+    authToken,
+    ctrlWrapper,
+    validateBody
+}
