+++ middlewares/validateBody.js
@@ -9,5 +9,5 @@ const validateBody = (schema) => {
     };
 };
   
-module.exports = { validateBody };
+module.exports = validateBody;
