+++ models/index.js
@@ -1,0 +1,7 @@
+const {Contact} = require("./contact")
+const {User} = require("./users");
+
+module.exports = {
+    Contact,
+    User
+}
