+++ models/users.js
@@ -1,7 +1,7 @@
 // const {Schema, model } = require("mongoose");
 const mongoose = require("mongoose");
 const bCrypt = require("bcryptjs");
-//const Joi = require("joi");
+const Joi = require("joi");
 
 const Schema = mongoose.Schema;
 
@@ -40,7 +40,7 @@ userSchema.methods.validPassword = function(password) {
     return bCrypt.compareSync(password, this.password);
 }
 
-/*const joiSignupSchema = Joi.object({
+const joiSignupSchema = Joi.object({
   username: Joi.string().required(),
   email: Joi.string().required(),
   password: Joi.string().min(6).required()
@@ -49,9 +49,12 @@ userSchema.methods.validPassword = function(password) {
 const joiLoginSchema = Joi.object({
   email: Joi.string().required(),
   password: Joi.string().min(6).required()
-})*/
+})
 
 const User = mongoose.model("user", userSchema);
 
-module.exports = User;
-
+module.exports = {
+  User,
+  joiLoginSchema,
+  joiSignupSchema
+}
